refac(crypto): extract `PoseidonConfigEntry(Base)` from `poseidon_config.h`
