+++ tachyon/crypto/hashes/sponge/poseidon/BUILD.bazel
@@ -17,11 +17,23 @@ tachyon_cc_library(
     name = "poseidon_config",
     hdrs = ["poseidon_config.h"],
     deps = [
-        ":poseidon_grain_lfsr",
+        ":poseidon_config_entry",
         "//tachyon/base/buffer:copyable",
         "//tachyon/base/ranges:algorithm",
         "@com_google_absl//absl/types:span",
     ],
+)
+
+tachyon_cc_library(
+    name = "poseidon_config_entry",
+    hdrs = ["poseidon_config_entry.h"],
+    deps = [":poseidon_config_entry_base"],
+)
+
+tachyon_cc_library(
+    name = "poseidon_config_entry_base",
+    hdrs = ["poseidon_config_entry_base.h"],
+    deps = [":poseidon_grain_lfsr"],
 )
 
 tachyon_cc_library(
