+++ tachyon/crypto/hashes/sponge/poseidon/poseidon_config.h
@@ -13,91 +13,10 @@
 #include "tachyon/base/buffer/copyable.h"
 #include "tachyon/base/logging.h"
 #include "tachyon/base/ranges/algorithm.h"
-#include "tachyon/crypto/hashes/sponge/poseidon/poseidon_grain_lfsr.h"
+#include "tachyon/crypto/hashes/sponge/poseidon/poseidon_config_entry.h"
 
 namespace tachyon {
 namespace crypto {
-
-template <typename PrimeField>
-struct PoseidonConfig;
-
-// An entry in the Poseidon config
-struct TACHYON_EXPORT PoseidonConfigEntry {
-  // The rate (in terms of number of field elements).
-  size_t rate;
-
-  // Exponent used in S-boxes.
-  uint64_t alpha;
-
-  // Number of rounds in a full-round operation.
-  size_t full_rounds;
-
-  // Number of rounds in a partial-round operation.
-  size_t partial_rounds;
-
-  // Number of matrices to skip when generating config using the Grain LFSR.
-  // The matrices being skipped are those that do not satisfy all the desired
-  // properties. See:
-  // https://extgit.iaik.tugraz.at/krypto/hadeshash/-/blob/master/code/generate_parameters_grain.sage
-  size_t skip_matrices;
-
-  constexpr PoseidonConfigEntry() : PoseidonConfigEntry(0, 0, 0, 0, 0) {}
-  constexpr PoseidonConfigEntry(size_t rate, uint64_t alpha, size_t full_rounds,
-                                size_t partial_rounds, size_t skip_matrices)
-      : rate(rate),
-        alpha(alpha),
-        full_rounds(full_rounds),
-        partial_rounds(partial_rounds),
-        skip_matrices(skip_matrices) {}
-
-  template <typename PrimeField>
-  PoseidonGrainLFSRConfig ToPoseidonGrainLFSRConfig() const {
-    PoseidonGrainLFSRConfig config;
-    config.prime_num_bits = PrimeField::kModulusBits;
-    config.state_len = rate + 1;
-    config.num_full_rounds = full_rounds;
-    config.num_partial_rounds = partial_rounds;
-    return config;
-  }
-
-  template <typename PrimeField>
-  PoseidonConfig<PrimeField> ToPoseidonConfig() const;
-
-  bool operator==(const PoseidonConfigEntry& other) const {
-    return rate == other.rate && alpha == other.alpha &&
-           full_rounds == other.full_rounds &&
-           partial_rounds == other.partial_rounds &&
-           skip_matrices == other.skip_matrices;
-  }
-  bool operator!=(const PoseidonConfigEntry& other) const {
-    return !operator==(other);
-  }
-};
-
-// An array of the default config optimized for constraints
-// (rate, alpha, full_rounds, partial_rounds, skip_matrices)
-// for rate = 2, 3, 4, 5, 6, 7, 8
-// Here, |skip_matrices| denotes how many matrices to skip before finding one
-// that satisfy all the requirements.
-constexpr const PoseidonConfigEntry kOptimizedConstraintsDefaultParams[] = {
-    PoseidonConfigEntry(2, 17, 8, 31, 0), PoseidonConfigEntry(3, 5, 8, 56, 0),
-    PoseidonConfigEntry(4, 5, 8, 56, 0),  PoseidonConfigEntry(5, 5, 8, 57, 0),
-    PoseidonConfigEntry(6, 5, 8, 57, 0),  PoseidonConfigEntry(7, 5, 8, 57, 0),
-    PoseidonConfigEntry(8, 5, 8, 57, 0),
-};
-
-// An array of the default config optimized for weights
-// (rate, alpha, full_rounds, partial_rounds, skip_matrices)
-// for rate = 2, 3, 4, 5, 6, 7, 8
-constexpr const PoseidonConfigEntry kOptimizedWeightsDefaultParams[] = {
-    PoseidonConfigEntry(2, 257, 8, 13, 0),
-    PoseidonConfigEntry(3, 257, 8, 13, 0),
-    PoseidonConfigEntry(4, 257, 8, 13, 0),
-    PoseidonConfigEntry(5, 257, 8, 13, 0),
-    PoseidonConfigEntry(6, 257, 8, 13, 0),
-    PoseidonConfigEntry(7, 257, 8, 13, 0),
-    PoseidonConfigEntry(8, 257, 8, 13, 0),
-};
 
 // ARK(AddRoundKey) is a matrix that contains an ARC(AddRoundConstant) array in
 // each row. Each constant is added to the |state| of each round of Poseidon.
